proper unit tests for `snip.js`
